+++ src/crypto/secret_box.cpp
@@ -9,15 +9,15 @@
 using namespace util::data;
 
 SecretBox::SecretBox(bytevector key) {
-    CRYPTO_REQUIRE(key.size() == crypto_secretbox_KEYBYTES, "provided key is too long or too short for SecretBox")
+    CRYPTO_REQUIRE(key.size() == KEY_LEN, "provided key is too long or too short for SecretBox")
 
     this->key = reinterpret_cast<byte*>(sodium_malloc(
-        crypto_secretbox_KEYBYTES
+        KEY_LEN
     ));
 
     CRYPTO_REQUIRE(this->key != nullptr, "sodium_malloc returned nullptr")
 
-    std::memcpy(this->key, key.data(), crypto_secretbox_KEYBYTES);
+    std::memcpy(this->key, key.data(), KEY_LEN);
 }
 
 SecretBox SecretBox::withPassword(const std::string_view pw) {
@@ -43,8 +43,10 @@ size_t SecretBox::encryptInto(const byte* src, byte* dest, size_t size) {
     byte nonce[NONCE_LEN];
     util::crypto::secureRandom(nonce, NONCE_LEN);
 
-    byte* ciphertext = dest + NONCE_LEN;
-    CRYPTO_ERR_CHECK(crypto_secretbox_easy(ciphertext, src, size, nonce, key), "crypto_secretbox_easy failed")
+    byte* mac = dest + NONCE_LEN;
+    byte* ciphertext = mac + MAC_LEN;
+
+    CRYPTO_ERR_CHECK(crypto_secretbox_xchacha20poly1305_detached(ciphertext, mac, src, size, nonce, key), "crypto_secretbox_detached failed")
 
     // prepend the nonce
     std::memcpy(dest, nonce, NONCE_LEN);
@@ -55,13 +57,13 @@ size_t SecretBox::encryptInto(const byte* src, byte* dest, size_t size) {
 size_t SecretBox::decryptInto(const byte* src, byte* dest, size_t size) {
     CRYPTO_REQUIRE(size >= PREFIX_LEN, "message is too short")
 
-    const byte* nonce = src;
-    const byte* ciphertext = src + NONCE_LEN;
-
     size_t plaintextLength = size - PREFIX_LEN;
-    size_t ciphertextLength = size - NONCE_LEN;
 
-    CRYPTO_ERR_CHECK(crypto_secretbox_open_easy(dest, ciphertext, ciphertextLength, nonce, key), "crypto_secretbox_open_easy failed")
+    const byte* nonce = src;
+    const byte* mac = src + NONCE_LEN;
+    const byte* ciphertext = mac + MAC_LEN;
+
+    CRYPTO_ERR_CHECK(crypto_secretbox_xchacha20poly1305_open_detached(dest, ciphertext, mac, plaintextLength, nonce, key), "crypto_secretbox_open_easy failed")
 
     return plaintextLength;
 }